% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/askgpt.R
\name{askgpt}
\alias{askgpt}
\title{Ask Chatgpt for an answer}
\usage{
askgpt(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{additional arguments}

\item{prompt}{idea}
}
\value{
code
}
\description{
Ask Chatgpt for an answer
}
