% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_comment.R
\name{code_comment}
\alias{code_comment}
\title{TidyChat: Code Comment}
\usage{
code_comment(x, ...)
}
\arguments{
\item{x}{A character vector, function, or fseq object.}

\item{...}{Additional arguments.}

\item{idea}{A character vector.}
}
\value{
A character vector.
}
\description{
TidyChat: Code Comment
}
