% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_completion.R
\name{code_completion}
\alias{code_completion}
\title{Code Completion}
\usage{
code_completion(x, ...)
}
\arguments{
\item{x}{A character vector, function, or fseq object.}

\item{...}{Additional arguments.}

\item{idea}{A character vector.}
}
\value{
A character vector.
}
\description{
Code Completion
}
