% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{set_tidychat}
\alias{set_tidychat}
\title{Set TidyChat Global Variables}
\usage{
set_tidychat(
  openai_api_key,
  proxy_url = NULL,
  proxy_port = NULL,
  openai_model = "gpt-3.5-turbo",
  openai_max_tokens = 256,
  openai_temperature = 1,
  openai_top_p = 1,
  openai_frequency_penalty = 0,
  openai_presence_penalty = 0,
  openai_verbose = TRUE,
  openai_return_language = "English"
)
}
\arguments{
\item{openai_api_key}{OpenAI API Key}

\item{proxy_url}{Proxy URL}

\item{proxy_port}{Proxy Port}

\item{openai_model}{OpenAI Model}

\item{openai_max_tokens}{OpenAI Max Tokens}

\item{openai_temperature}{OpenAI Temperature}

\item{openai_top_p}{OpenAI Top P}

\item{openai_frequency_penalty}{OpenAI Frequency Penalty}

\item{openai_presence_penalty}{OpenAI Presence Penalty}

\item{openai_verbose}{OpenAI Verbose}

\item{openai_return_language}{OpenAI Return Language}
}
\description{
Set TidyChat Global Variables
}
