% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{update_tidychat}
\alias{update_tidychat}
\title{Update TidyChat Global Variables}
\usage{
update_tidychat(key, value)
}
\arguments{
\item{key}{Key}

\item{value}{Value}
}
\description{
Update TidyChat Global Variables
}
